- fix mempool warning - weak ptr redundant checks removed - ppmem changes memory mark on freed objects (to not go through same dealloc twice and crash on libc free)

+++ public/ds/mempool.h
@@ -79,7 +79,7 @@ private:
 		Block*			next{ nullptr };
 	};
 
-	static constexpr const int blockSize = max(sizeof(T), sizeof(Block));
+	static constexpr const size_t blockSize = max(sizeof(T), sizeof(Block));
 
 	struct Buffer
 	{
